safari-600.7-branch/LayoutTests/fast/dom/custom/lifecycle-ready-innerHTML-expected.txt

This test ensures that the lifecycle callback of a parser-made element is visible in following script block.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.callbacksCalled is ['a', 'b', 'c']
PASS window.callbacksCalled is ['a', 'b', 'c']
PASS successfullyParsed is true

TEST COMPLETE
